feat(auth): refactored authorization without test
