+++ backend/server/schedule_get.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, jsonify, make_response
+from server.auth import auth
 from tools.get_slots_by_filter import get_slots_by_filter
 from tools.get_slots_from_db_for_schedule import get_slots_from_db_for_schedule
 
@@ -6,6 +7,7 @@ schedule_get = Blueprint('schedule_get', __name__)
 
 
 @schedule_get.route("/schedule/status=<filter>", methods=['GET'])
+@auth.login_required
 def slot_by_filter(filter):
     """ Filter: booked or available.
         Return JSON with <filter> slots.
@@ -15,6 +17,7 @@ def slot_by_filter(filter):
 
 @schedule_get.route("/schedule/day=<date>/", defaults={'filter': None}, methods=['GET'])
 @schedule_get.route("/schedule/day=<date>&status=<filter>", methods=['GET'])
+@auth.login_required
 def slot_in_day_by_filter(date, filter):
     """ Filter: booked or available.
         Return JSON of <filter> slots for the <date>
